Move GraphQL to a separate package.
