+++ Cargo.toml
@@ -39,6 +39,7 @@ route-recognizer = "0.3.0"
 cynic = { version = "0.14" }
 itertools = "0.10.2"
 tower = { version = "0.4.13", features = ["util", "limit", "buffer", "load-shed"] }
+github-graphql = { path = "github-graphql" }
 
 [dependencies.serde]
 version = "1"
