+++ github-graphql/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "github-graphql"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+cynic = { version = "0.14" }
